+++ include/parser.h
@@ -58,13 +58,20 @@ typedef enum PercyMemoryMagnitude MemMag;
 
 extern const complex CMPLX_MIN;
 extern const complex CMPLX_MAX;
+extern const long double complex LCMPLX_MIN;
+extern const long double complex LCMPLX_MAX;
 
 
 ParseErr stringToULong(unsigned long *x, char *nptr, unsigned long min, unsigned long max, char **endptr, int base);
 ParseErr stringToUIntMax(uintmax_t *x, char *nptr, uintmax_t min, uintmax_t max, char **endptr, int base);
 ParseErr stringToDouble(double *x, char *nptr, double min, double max, char **endptr);
+ParseErr stringToDoubleL(long double *x, char *nptr, long double min, long double max, char **endptr);
 ParseErr stringToComplexPart(complex *z, char *nptr, complex min, complex max, char **endptr, ComplexPt *type);
+ParseErr stringToComplexPartL(long double complex *z, char *nptr, long double complex min, long double complex max,
+                                char **endptr, ComplexPt *type);
 ParseErr stringToComplex(complex *z, char *nptr, complex min, complex max, char **endptr);
+ParseErr stringToComplexL(long double complex *z, char *nptr, long double complex min, long double complex max,
+                             char **endptr);
 ParseErr stringToMemory(size_t *bytes, char *nptr, size_t min, size_t max, char **endptr, int magnitude);
 
 size_t strncpyGraph(char *dest, const char *src, size_t n);